Rename conflicting files
We already have config.h and internal.h at the toplevel, and it's
confusing to have another, as compilation then depends on `-I` search
path.
