+++ gg/lib/ggconfig.h
