+++ zephyr/ZCmpUIDP.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 int
 ZCompareUIDPred(ZNotice_t *notice, void *uid)
