+++ zephyr/ZFmtList.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 Code_t
 ZFormatNoticeList(ZNotice_t *notice, register char **list, int nitems,
