+++ zephyr/ZFmtSmRaw.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 Code_t
 ZFormatSmallRawNotice(ZNotice_t *notice, ZPacket_t buffer, int *ret_len)
