+++ zephyr/ZGetLocs.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 #define min(a,b) ((a)<(b)?(a):(b))
