+++ zephyr/ZGetSender.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 #ifndef WIN32
 #include <pwd.h>
