+++ zephyr/ZGetWGPort.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 int
 ZGetWGPort(void)
