+++ zephyr/ZIfNotice.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 Code_t
 ZIfNotice(ZNotice_t *notice, struct sockaddr_in *from,
