+++ zephyr/ZLocations.c
@@ -9,7 +9,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 #include "util.h"
 
 #ifndef WIN32
