+++ zephyr/ZPeekNot.c
@@ -8,7 +8,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 Code_t
 ZPeekNotice(ZNotice_t *notice, struct sockaddr_in *from)
