+++ zephyr/ZWait4Not.c
@@ -7,7 +7,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 
 #ifdef WIN32
 #include <winsock2.h>
