+++ zephyr/Zinternal.c
@@ -9,7 +9,7 @@
  *	"mit-copyright.h".
  */
 
-#include "internal.h"
+#include "zephyrlib_internal.h"
 #ifdef WIN32
 #include <winsock2.h>
